fixup! Add script to adapt data in database to current Concent version
